Moved DT dataprocessing function to a new .py file
